25.09.17 骑士团加深关系补全 骑士团主动加深关系补全 骑士团主动拜访补全 关系工具函数独立 新增普通骑士团信件类 骑士团协助类提升为骑士团交互类 获取玩家基地Map同步奥德赛机制 文件夹结构优化
25.09.16
骑士拜访实现
添加交互总记录

+++ Components/TagStrToValue/TagStrToFloat.cs
@@ -5,7 +5,7 @@ namespace OberoniaAurea.RatkinOrder;
 public class TagStrToFloat : TagStrToValue<float>
 {
     public TagStrToFloat() : base() { }
-    public TagStrToFloat(int defaultValue, bool removeWhenDefault) : base(defaultValue, removeWhenDefault, LookMode.Value) { }
+    public TagStrToFloat(float defaultValue, bool removeWhenDefault) : base(defaultValue, removeWhenDefault, LookMode.Value) { }
 
     public override void OffsetTagValueBy(string tag, float offset, bool addIfMiss)
     {
